requirements.txt
requests==2.21.0
Flask==1.0.2
mysql_connector


